Conflicts:
	core/src/com/mygdx/game/Character.java
	core/src/com/mygdx/game/Level.java
	core/src/com/mygdx/game/LevelOne.java
	core/src/com/mygdx/game/Platform.java
	core/src/com/mygdx/game/testFalling.java
	desktop/src/com/mygdx/game/desktop/DesktopLauncher.java

+++ core/src/com/mygdx/game/Fireboy.java
@@ -22,7 +22,7 @@ public class Fireboy extends Character {
      * @param x a float representing the x position on the screen
      * @param y a float representing the y position on the screen
      */
-    public Fireboy(int x, int y) {
+    public Fireboy(float x, float y) {
         super(x, y);
     }
 
@@ -34,14 +34,9 @@ public class Fireboy extends Character {
      */
     @Override
     public void draw(ShapeRenderer shapeBatch) {
-   //      set the Fireboy to be red
+        // set the Fireboy to be red
         shapeBatch.setColor(Color.RED);
-     //    draw the Fireboy on the screen
+        // draw the Fireboy on the screen
         super.draw(shapeBatch);
     }
-    
- 
-    }
-    
-
-
+}
